+++ azdev/commands.py
@@ -67,6 +67,7 @@ def load_command_table(self, _):
         g.command('publish', 'publish_extensions')
         g.command('update-index', 'update_extension_index')
         g.command('cal-next-version', 'cal_next_version')
+        g.command('show', 'show_extension')
 
     with CommandGroup(self, 'extension repo', operation_group('extensions')) as g:
         g.command('add', 'add_extension_repo')